Create README.md
I am an FPV flying enthusiast, but after years of flying, I feel a little tired of it, so I thought about doing something. I thought this kind of FPV flying machine would be very interesting when it flies, so I drew a 3D design drawing of it. This FPV flying machine will use duct as power, and thanks to its powerful power and this aerodynamic shape, it should have a very high speed when it flies.